pyproject.toml
[project]
name = "founderpilot"
version = "0.1.0"
description = "Suite of autonomous agents for founders: InboxPilot, InvoicePilot, MeetingPilot"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "FounderPilot Team"}
]

dependencies = [
    # Core
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",

    # Database
    "sqlalchemy[asyncio]>=2.0.25",
    "asyncpg>=0.29.0",
    "alembic>=1.13.0",

    # Redis & Celery
    "redis>=5.0.0",
    "celery[redis]>=5.3.0",

    # Auth
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "httpx>=0.26.0",

    # Google APIs
    "google-api-python-client>=2.100.0",
    "google-auth-oauthlib>=1.0.0",
    "google-auth-httplib2>=0.2.0",

    # LangGraph & LLM
    "langgraph>=0.0.40",
    "langgraph-checkpoint-postgres>=0.0.5",
    "langchain>=0.1.0",
    "langchain-openai>=0.0.5",
    "langchain-anthropic>=0.1.0",

    # Slack
    "slack-bolt>=1.18.0",

    # Observability
    "langfuse>=2.0.0",
    "structlog>=24.1.0",

    # Payments
    "stripe>=7.0.0",

    # Utils
    "python-multipart>=0.0.6",
    "cryptography>=42.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "httpx>=0.26.0",
    "ruff>=0.1.0",
    "mypy>=1.8.0",
    "testcontainers>=3.7.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.mypy]
python_version = "3.11"
strict = true
ignore_missing_imports = true
